fix decoder & tests
