+++ reports/report_measurement_unit.rtf
@@ -13,10 +13,10 @@
 \paperw11907\paperh16838\margt1000\margl1200\margb1000\margr1200\sectd{\header}
 {\footer}
 {\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 base_measure_unit\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 name\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 uid\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 unit\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 г\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 9a4ec81b-df59-4b14-8bb5-74f0723216bf\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 9a4ec81b-df59-4b14-8bb5-74f0723216bf\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 кг\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 6461d14e-2b1a-45f5-b34a-bfbe9ae2faf4\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1000.0\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 6461d14e-2b1a-45f5-b34a-bfbe9ae2faf4\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 т\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 b68234fe-a305-43ee-9d78-9109658113ae\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1000.0\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 шт\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 770a1f07-2636-4163-8654-395b01add543\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 ч.л.\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 ed8299a8-c607-4c74-9245-09f108e516ad\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 мл\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 677a7c54-1a5b-4d4e-a6d5-150bfd9443b2\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 г\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 6fa893f9-a9c5-4467-9852-72685a66fd1c\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 6fa893f9-a9c5-4467-9852-72685a66fd1c\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 кг\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 291c63f2-f2ce-45cd-a885-f86f4403810b\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1000.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 291c63f2-f2ce-45cd-a885-f86f4403810b\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 т\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 fe40e353-0d74-443f-b15c-fc936d565124\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1000.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 шт\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 f1a442d9-fcc2-4628-9d0e-814c3cf51057\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 ч.л.\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 63df2231-f327-4cba-8124-c868fde970a0\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 мл\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 7594ad19-af5e-485f-968f-de895da4504b\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 1.0\cell\row}
 }
