+++ reports/report_nomenclature.csv
@@ -1,3 +1,3 @@
 full_name;measurement_unit;name;nomenclature_group;uid;
-None;1baa18f2-0c57-4409-8320-6237f7f0bd51;Мука;1ff2de59-6e3b-4cc2-9bff-250059f06422;b45c10c9-230a-4ead-9748-aa0079db8fa2;
-None;2a1bb569-0951-429b-a907-6f22e148bf70;Лёд;10e23948-5768-44af-8fdf-d096ce372fd3;1c96d4a2-43b9-4f91-a7d1-61f1b2781d2b;
+None;b09a5a66-50b8-4a68-9d73-6dd18758c926;Мука;ad5d5ad3-768d-477f-bf78-69023ba54968;28153d0d-2d28-4909-b267-d946b9c6e62e;
+None;b3304025-00d1-4cb9-8302-ec5809b9c194;Лёд;ef91ad2a-6c5c-48f6-a9d0-d56d97cb5bde;0e289f0b-136d-4f62-bfed-afaf1f7e9ab0;
