+++ reports/report_nomenclature.rtf
@@ -13,6 +13,6 @@
 \paperw11907\paperh16838\margt1000\margl1200\margb1000\margr1200\sectd{\header}
 {\footer}
 {\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\cellx10500\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 full_name\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 measurement_unit\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 name\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 nomenclature_group\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 uid\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\cellx10500\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 827e4d02-a0e9-4c2e-ac2e-6b1fd2fc01de\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 Мука\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 dadec6f7-8c39-4c82-9bac-2b2ca8444505\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 dc4a9c5f-055a-41bf-9614-5ba580f56638\cell\row}
-{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\cellx10500\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 8db1b278-5412-4a99-8161-56a318742cbc\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 Лёд\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 55d5c251-dd35-429d-86b1-2ea3374ba66f\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 f3664422-f1a3-4319-8dd3-549702916c1c\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\cellx10500\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 083ad8cd-23cd-4cfd-ae97-fdebe4445c27\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 Мука\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 7575b084-b49a-45ce-834a-bbfa73d39c03\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 3cbe29cb-fee8-41ee-a8a2-95e34583314e\cell\row}
+{\trowd\trgaph108\trql\cellx2100\cellx4200\cellx6300\cellx8400\cellx10500\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 None\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 a7d21f68-2813-42ae-805e-1f8d532f36a6\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 Лёд\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 fdde982e-ad2d-4f64-9477-03ed8901984c\cell\pard\plain\intbl \s0\ql\sb60\sa60\f0\fs22 8acef339-f637-4d68-882a-04e9b220878c\cell\row}
 }
