--- src/utils/decoders/ingridient_json_decoder.py
@@ -1,2 +1,0 @@
-class IngridientJsonDecoder:
-    pass
